protect from deleting and editing by unauthorized users

+++ rest-api/controllers/feed.js
@@ -95,6 +95,9 @@ export const updatePost = async (req, res, next) => {
         if (!post) {
             return next(new HttpError(404, 'No post found!'));
         }
+        if(post.creator.toString() !== req.userId) {
+            return next(new HttpError(403, `Not authorized! ${req.userId} is not owner of this post`));
+        }
         if (imageUrl !== post.imageUrl) {
             await deleteFile(post.imageUrl);
         }
@@ -122,6 +125,9 @@ export const deletePost = async (req, res, next) => {
         const post = await Post.findById(postId);
         if (!post) {
             return next(new HttpError(404, 'Could not find post!'));
+        }
+        if(post.creator.toString() !== req.userId) {
+            return next(new HttpError(403, `Not authorized! ${req.userId} is not owner of this post`));
         }
         await deleteFile(post.imageUrl);
         await Post.deleteOne({ _id: postId });
